+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   [#778](https://github.com/nextcloud/cookbook/pull/778) @christianlupus
 - Usage of PHAR-based PHPUnit to avoid dependency on nikic/php-parser and dependency conflicts
   [#780](https://github.com/nextcloud/cookbook/pull/780) @christianlupus
+- Extracted abstract class for migration testing
+  [#783](https://github.com/nextcloud/cookbook/pull/783) @christianlupus
 
 ### Removed
 - Removed app info XML file to avoid confusion